Auto-Coder: Address issue #462
