+++ pyproject.toml
@@ -203,4 +203,5 @@ dev = [
     "black>=25.9.0",
     "isort>=5.0.0",
     "flake8>=6.0.0",
+    "types-toml>=0.10.8.20240310",
 ]
