+++ src/auto_coder/cli_commands_utils.py
@@ -8,8 +8,8 @@ import click
 from .auth_utils import get_auth_status, get_github_token
 from .automation_config import AutomationConfig
 from .automation_engine import AutomationEngine
+from .git_branch import migrate_pr_branches
 from .git_info import get_current_repo_name, is_git_repository
-from .git_utils import migrate_pr_branches
 from .github_client import GitHubClient
 from .logger_config import setup_logger
 from .util.github_action import get_github_actions_logs_from_url
